chore: move Shift+Enter keyboard handling to AbstractThreePanelScreen
It was duplicated across several subclasses, but this is standard
behaviour
