+++ common/src/main/java/dev/ftb/mods/ftblibrary/config/ui/EditConfigScreen.java
@@ -184,9 +184,6 @@ public class EditConfigScreen extends AbstractThreePanelScreen<EditConfigScreen.
     public boolean keyPressed(Key key) {
         if (super.keyPressed(key)) {
             return true;
-        } else if ((key.is(InputConstants.KEY_RETURN) || key.is(InputConstants.KEY_NUMPADENTER)) && key.modifiers.shift()) {
-            doAccept();
-            return true;
         } else if (key.is(InputConstants.KEY_ADD) || key.is(InputConstants.KEY_EQUALS)) {
             buttonExpandAll.onClicked(MouseButton.LEFT);
         } else if (key.is(InputConstants.KEY_MINUS) || key.is(GLFW.GLFW_KEY_KP_SUBTRACT)) {
