+++ common/src/main/java/dev/ftb/mods/ftblibrary/config/ui/ResourceSelectorScreen.java
@@ -160,18 +160,6 @@ public abstract class ResourceSelectorScreen<T> extends AbstractThreePanelScreen
         }
     }
 
-    @Override
-    public boolean keyPressed(Key key) {
-        if (super.keyPressed(key)) {
-            return true;
-        } else if ((key.is(InputConstants.KEY_RETURN) || key.is(InputConstants.KEY_NUMPADENTER)) && key.modifiers.shift()) {
-            doAccept();
-            return true;
-        }
-
-        return false;
-    }
-
     protected void setSelected(SelectableResource<T> stack) {
         long count = selectedStack == null || selectedStack.isEmpty() ? Math.max(stack.getCount(), countBox.getCount()) : selectedStack.getCount();
         selectedStack = stack.copyWithCount(count);
